DType Management (#269)
* WIP

* fix broken tests

* Stub test cases

* Fix warnings

* Update tests

* spelling

* Update test/unit/test_data.py

* revert timestamps 32

* clean

* Verify no nans

+++ test/unit/test_hooks/test_deduplication_hook.py
@@ -8,8 +8,8 @@ from tgm.hooks import DeduplicationHook
 
 @pytest.fixture
 def dg():
-    edge_index = torch.LongTensor([[2, 2], [2, 4], [1, 8]])
-    edge_timestamps = torch.LongTensor([1, 5, 20])
+    edge_index = torch.IntTensor([[2, 2], [2, 4], [1, 8]])
+    edge_timestamps = torch.IntTensor([1, 5, 20])
     data = DGData.from_raw(edge_timestamps, edge_index)
     return DGraph(data)
 
@@ -28,33 +28,33 @@ def test_dedup(dg):
     batch = dg.materialize()
     processed_batch = hook(dg, batch)
     torch.testing.assert_close(
-        processed_batch.unique_nids, torch.LongTensor([1, 2, 4, 8])
+        processed_batch.unique_nids, torch.IntTensor([1, 2, 4, 8])
     )
     torch.testing.assert_close(
-        processed_batch.global_to_local(batch.src), torch.LongTensor([1, 1, 0])
+        processed_batch.global_to_local(batch.src), torch.IntTensor([1, 1, 0])
     )
     torch.testing.assert_close(
-        processed_batch.global_to_local(batch.dst), torch.LongTensor([1, 2, 3])
+        processed_batch.global_to_local(batch.dst), torch.IntTensor([1, 2, 3])
     )
 
 
 def test_dedup_with_negatives(dg):
     hook = DeduplicationHook()
     batch = dg.materialize()
-    batch.neg = torch.LongTensor([1, 5, 10])  # add some mock negatives
+    batch.neg = torch.IntTensor([1, 5, 10])  # add some mock negatives
 
     processed_batch = hook(dg, batch)
     torch.testing.assert_close(
-        processed_batch.unique_nids, torch.LongTensor([1, 2, 4, 5, 8, 10])
+        processed_batch.unique_nids, torch.IntTensor([1, 2, 4, 5, 8, 10])
     )
     torch.testing.assert_close(
-        processed_batch.global_to_local(batch.src), torch.LongTensor([1, 1, 0])
+        processed_batch.global_to_local(batch.src), torch.IntTensor([1, 1, 0])
     )
     torch.testing.assert_close(
-        processed_batch.global_to_local(batch.dst), torch.LongTensor([1, 2, 4])
+        processed_batch.global_to_local(batch.dst), torch.IntTensor([1, 2, 4])
     )
     torch.testing.assert_close(
-        processed_batch.global_to_local(batch.neg), torch.LongTensor([0, 3, 5])
+        processed_batch.global_to_local(batch.neg), torch.IntTensor([0, 3, 5])
     )
 
 
@@ -62,24 +62,24 @@ def test_dedup_with_nbrs(dg):
     hook = DeduplicationHook()
     batch = dg.materialize()
     batch.nbr_nids = [  # add some mock neighbours
-        torch.LongTensor([1, 5]),  # First hop
-        torch.LongTensor([10]),  # Second hop
+        torch.IntTensor([1, 5]),  # First hop
+        torch.IntTensor([10]),  # Second hop
     ]
-    batch.nbr_mask = [torch.LongTensor([1, 1]), torch.LongTensor([1])]
+    batch.nbr_mask = [torch.IntTensor([1, 1]), torch.IntTensor([1])]
 
     processed_batch = hook(dg, batch)
     torch.testing.assert_close(
-        processed_batch.unique_nids, torch.LongTensor([1, 2, 4, 5, 8, 10])
+        processed_batch.unique_nids, torch.IntTensor([1, 2, 4, 5, 8, 10])
     )
     torch.testing.assert_close(
-        processed_batch.global_to_local(batch.src), torch.LongTensor([1, 1, 0])
+        processed_batch.global_to_local(batch.src), torch.IntTensor([1, 1, 0])
     )
     torch.testing.assert_close(
-        processed_batch.global_to_local(batch.dst), torch.LongTensor([1, 2, 4])
+        processed_batch.global_to_local(batch.dst), torch.IntTensor([1, 2, 4])
     )
     torch.testing.assert_close(
-        processed_batch.global_to_local(batch.nbr_nids[0]), torch.LongTensor([0, 3])
+        processed_batch.global_to_local(batch.nbr_nids[0]), torch.IntTensor([0, 3])
     )
     torch.testing.assert_close(
-        processed_batch.global_to_local(batch.nbr_nids[1]), torch.LongTensor([5])
+        processed_batch.global_to_local(batch.nbr_nids[1]), torch.IntTensor([5])
     )
